feat: CRUD tenants

+++ frontend/src/pages/Residents/index.js
@@ -1,5 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
+import { toast } from 'react-toastify';
+import { format } from 'date-fns';
+import pt from 'date-fns/locale/pt';
 import {
   MdModeEdit,
   MdDelete,
@@ -16,23 +19,21 @@ function Residents() {
   const { id } = useParams();
 
   async function deleteResident(resident_id) {
-    await api.delete(`tenants/${resident_id}`);
-    history.go(0);
+    if (residents.length > 1) {
+      await api.delete(`tenants/${resident_id}`);
+      history.go(0);
+    } else {
+      toast.error('O apartamento precisa de no mínimo 1 morador.');
+    }
   }
 
   useEffect(() => {
     async function loadResidents() {
-      const response = await api.get(`tenants/${id}`);
-
-      const data = response.data.map((resident) => {
-        return {
-          name: resident.name,
-          email: resident.email,
-          id: resident.id,
-        };
+      const response = await api.get(`tenants`, {
+        params: { apartment_id: id },
       });
 
-      setResidents(data);
+      setResidents(response.data);
     }
 
     loadResidents();
@@ -54,21 +55,42 @@ function Residents() {
         {residents.map((resident) => (
           <Resident key={resident.id}>
             <strong>{resident.name}</strong>
-            <span>{resident.email}</span>
-            <button type="button" onClick={() => {}}>
+            <p>{resident.email}</p>
+            <p>
+              {format(
+                new Date(resident.birtday),
+                "'Aniversário ' d 'de' MMMM",
+                {
+                  locale: pt,
+                }
+              )}
+            </p>
+            <p>{`Telefone: ${resident.phone}`}</p>
+            <p>{`CPF: ${resident.cpf}`}</p>
+            {resident.accountable && <span>Responsável</span>}
+            <button
+              type="button"
+              onClick={() => {
+                history.push(`/resident-edit/${resident.id}`);
+              }}
+            >
               <MdModeEdit size={24} />
             </button>
             <button
               type="button"
               onClick={() => {
-                deleteResident(resident.id);
+                if (resident.accountable) {
+                  toast.error('Para excluir outro deve ser responsável.');
+                } else {
+                  deleteResident(resident.id);
+                }
               }}
             >
               <MdDelete size={24} />
             </button>
-            <button type="button" onClick={() => {}}>
-              <MdKeyboardArrowRight size={24} />
-            </button>
+            {/* <button type="button" onClick={() => {}}>
+              Tornar responsável
+            </button> */}
           </Resident>
         ))}
       </ul>
